Don't bitshift player yaw and pitch input

+++ src/d_protocol.h
@@ -69,9 +69,9 @@ struct usercmd_t
 	// [BB] Added unused dummy variables to prevent sthook from working at all.
 	BYTE	dummy[42];
 	DWORD	buttons;
-	short	pitch;			// up/down
-	short	yaw;			// left/right
-	short	roll;			// "tilt"
+	long	pitch;			// up/down
+	long	yaw;			// left/right
+	long	roll;			// "tilt"
 	short	forwardmove;
 	short	sidemove;
 	short	upmove;
